requirements.txt
paho-mqtt==1.3.1
pkg-resources==0.0.0
pyserial==3.4
six==1.11.0
